django_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow>=10.2.0
python-decouple==3.8
psycopg2==2.9.7
django-filter==23.5
django-ratelimit==4.1.0